clients: de-dupe async_query fn in npms and npm reg clients rename it to request_json
