+++ depobs/clients/npmsio.py
@@ -2,24 +2,15 @@ import asyncio
 import logging
 from typing import Any, AsyncGenerator, Dict, Iterable, Optional
 
-import aiohttp
-
-from depobs.clients.aiohttp_client import AIOHTTPClientConfig, aiohttp_session
+from depobs.clients.aiohttp_client import (
+    AIOHTTPClientConfig,
+    aiohttp_session,
+    request_json,
+)
 from depobs.util.serialize_util import grouper
 from depobs.util.type_util import Result
 
 log = logging.getLogger(__name__)
-
-
-async def async_query(
-    session: aiohttp.ClientSession, url: str, json: Iterable[str]
-) -> Optional[Dict]:
-    log.debug(f"posting {json} to {url}")
-    response_json: Optional[Dict] = None
-    response = await session.post(url, json=json)
-    response_json = await response.json()
-    log.debug(f"got response json {response_json!r}")
-    return response_json
 
 
 async def fetch_npmsio_scores(
@@ -35,10 +26,11 @@ async def fetch_npmsio_scores(
     async with aiohttp_session(config) as s:
         group_results = await asyncio.gather(
             *[
-                async_query(
+                request_json(
                     s,
+                    "POST",
                     f"{config['base_url']}package/mget",
-                    [
+                    json=[
                         package_name
                         for package_name in group
                         if package_name is not None
@@ -52,7 +44,13 @@ async def fetch_npmsio_scores(
         # pull {data1}, {data2} from {package_name_1: {data1}, package_name_2: {data2}}
         for group_result in group_results:
             if group_result is None:
+                log.warn(f"got None npms.io group for package_names {package_names}")
                 continue
 
             for result in group_result.values():
+                if result is None:
+                    log.warn(
+                        f"got None npms.io results for package_names {package_names}"
+                    )
+                    continue
                 yield result
